+++ lib/model/http/logout_pbx_res.dart
@@ -1,0 +1,17 @@
+class LogoutPbxRes {
+  String? message;
+
+  LogoutPbxRes({
+    this.message,
+  });
+
+  LogoutPbxRes.fromJson(Map<String, dynamic> json) {
+    message = json['message'] as String?;
+  }
+
+  Map<String, dynamic> toJson() {
+    final Map<String, dynamic> json = <String, dynamic>{};
+    json['message'] = message;
+    return json;
+  }
+}