+++ lib/services/pitel_service_interface.dart
@@ -1,3 +1,4 @@
+import 'package:flutter_pitel_voip/model/http/logout_pbx_res.dart';
 import 'package:flutter_pitel_voip/sip/src/sip_ua_helper.dart';
 
 import 'models/pn_push_params.dart';
@@ -12,4 +13,9 @@ abstract class PitelService {
   Future<PitelSettings> registerSipWithoutFCM(
     PnPushParams pnPushParams,
   );
+  Future<LogoutPbxRes?> logoutPbx({
+    required String extension,
+    required String authorization,
+    required String apiUrl,
+  });
 }